refactoring, add predator bear and method eat for predators, moved enum to a individual file, test for concurrent

+++ src/main/java/com/komarov/lost/floraAndFauna/Eateble.java
@@ -1,10 +1,8 @@
 package com.komarov.lost.floraAndFauna;
 
 public interface Eateble {
-
-    default int getCaloric(){
+    default double getCaloric(){
         return getWeight();
     }
-
-    int getWeight();
+    double getWeight();
 }
