+++ src/main/java/com/komarov/lost/floraAndFauna/plants/Plant.java
@@ -1,23 +1,20 @@
 package com.komarov.lost.floraAndFauna.plants;
-
 import com.komarov.lost.floraAndFauna.Eateble;
 import lombok.Getter;
-
 
 public class Plant implements Eateble {
     private final static String PLANT_EMOJI = "\uD83C\uDF3F"; // 🌿
     @Getter
     private static final int MAX_POPULATION_ON_AREA = 200;
     @Getter
-    private static final int WEIGHT = 1;
+    private static final double WEIGHT = 1.0d;
 
     @Override
     public String toString() {
         return PLANT_EMOJI;
     }
-
     @Override
-    public int getWeight() {
+    public double getWeight() {
         return WEIGHT;
     }
 }
